pyproject.toml
[project]
name = "buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.18",
    "fastapi[standard]>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "google-cloud-speech>=2.32.0",
    "google-cloud-texttospeech>=2.27.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "psycopg>=3.2.9",
    "pydub>=0.25.1",
    "sentry-sdk[fastapi]>=2.30.0",
    "tortoise-orm>=0.25.0",
    "websockets==10.1",
]
